memperbarui transaksi dan item

+++ database/migrations/2024_10_11_073322_add_category_id_to_items_table.php
@@ -11,8 +11,11 @@ return new class extends Migration
     public function up()
 {
     Schema::table('items', function (Blueprint $table) {
-        $table->dropColumn('kategori');
+        if (Schema::hasColumn('items', 'kategori')) {
+            $table->dropColumn('kategori');
+        }
     });
+    
 }
 
 public function down()
